Refactor for PR as proposed on:
https://github.com/AmpersandJS/ampersand-dom-bindings/pull/54#discussion_r90333447

to make a cleaner/more readable code. firstMatchOnly is now accepted (and invoked)
on getMatches method as a third parameter stating if the function should
return only the first matched element or all.

NOTE: it could become an optional func passed to array.filter to build custom
filtered results, but I could be just overthinking this.